Remove desktop token api
